add unit tests for command manager get rc override function

+++ include/comm_link.h
@@ -156,7 +156,7 @@ public:
   virtual void send_sonar(uint8_t system_id,
                           /* TODO enum type*/ uint8_t type, float range, float max_range,
                           float min_range) = 0;
-  virtual void send_status(uint8_t system_id, bool armed, bool failsafe, uint8_t rc_override,
+  virtual void send_status(uint8_t system_id, bool armed, bool failsafe, uint16_t rc_override,
                            bool offboard, uint8_t error_code, uint8_t control_mode,
                            int16_t num_errors, int16_t loop_time_us) = 0;
   virtual void send_timesync(uint8_t system_id, int64_t tc1, int64_t ts1) = 0;
